Save Schedule MVVM

+++ wpfzoo/viewmodel/ScheduleAdminVM.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
+using System.Windows.Controls;
 using wpfzoo.database;
 using wpfzoo.entities;
 using wpfzoo.views.administration;
@@ -15,12 +17,16 @@ namespace wpfzoo.viewmodel
         private Schedule currentSchedule;
         private ScheduleAdmin scheduleAdmin;
         MySQLManager<Schedule> scheduleManager = new MySQLManager<Schedule>();
+        ObservableCollection<Schedule> scheduleList = new ObservableCollection<Schedule>();
+
 
         public ScheduleAdminVM(ScheduleAdmin scheduleAdmin)
         {
             this.scheduleAdmin = scheduleAdmin;
+            this.scheduleAdmin.listScheduleUC.ItemsList.SelectionChanged += ItemsList_SelectionChanged;
 
             InitUC();
+            InitLists();
             InitActions();
         }
 
@@ -30,26 +36,51 @@ namespace wpfzoo.viewmodel
             this.scheduleAdmin.scheduleUC.Schedule = currentSchedule;
         }
 
+        private async void InitLists()
+        {
+            this.scheduleAdmin.listScheduleUC.LoadItem((await scheduleManager.Get()).ToList());
+        }
+
         private void InitActions()
         {
-            this.scheduleAdmin.btnDelete.Click += btnDeleteClick;
-            this.scheduleAdmin.btnOk.Click += btnOkClick;
-            this.scheduleAdmin.btnNew.Click += btnNewClick;
+            this.scheduleAdmin.btnDelete.Click += btnDelete_Click;
+            this.scheduleAdmin.btnOk.Click += btnOk_Click;
+            this.scheduleAdmin.btnNew.Click += btnNew_Click;
         }
 
-        private void btnDeleteClick(object sender, RoutedEventArgs e)
+        private async void btnDelete_Click(object sender, RoutedEventArgs e)
         {
-            scheduleManager.Delete(this.scheduleAdmin.scheduleUC.Schedule);
+            if (this.scheduleAdmin.scheduleUC.Schedule.Id != 0)
+            {
+                await scheduleManager.Delete(this.scheduleAdmin.scheduleUC.Schedule);
+                InitLists();
+            }
         }
 
-        private void btnOkClick(object sender, RoutedEventArgs e)
+        private async void btnOk_Click(object sender, RoutedEventArgs e)
         {
-            scheduleManager.Update(this.scheduleAdmin.scheduleUC.Schedule);
+            if (this.scheduleAdmin.scheduleUC.Schedule.Id != 0)
+            {
+                await scheduleManager.Update(this.scheduleAdmin.scheduleUC.Schedule);
+            }
+            else
+            {
+                await scheduleManager.Insert(this.scheduleAdmin.scheduleUC.Schedule);
+            }
         }
 
-        private void btnNewClick(object sender, RoutedEventArgs e)
+        private async void btnNew_Click(object sender, RoutedEventArgs e)
         {
-            scheduleManager.Insert(this.scheduleAdmin.scheduleUC.Schedule);
+            await scheduleManager.Insert(this.scheduleAdmin.scheduleUC.Schedule);
+        }
+
+        private void ItemsList_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            if (e.AddedItems.Count > 0)
+            {
+                Schedule item = (e.AddedItems[0] as Schedule);
+                this.scheduleAdmin.scheduleUC.Schedule = item;
+            }
         }
     }
 }
